+++ Akinator/Models/DB/User.cs
@@ -4,7 +4,7 @@
     [PetaPoco.PrimaryKey("user_id")]
     public class User
     {
-        public long user_id { get; set; }
+        public int user_id { get; set; }
         public bool is_doctor { get; set; }
         public string login { get; set; }
     }
